3:Integration Test

+++ src/main/java/com/demo/awsdemo/service/DemoService.java
@@ -11,10 +11,11 @@ import java.util.Optional;
 @Component
 public class DemoService {
 
-//    DemoDAO demoDao = new DemoDAO();
-    @Autowired
-    DemoDAO demoDao;
+    DemoDAO demoDao = new DemoDAO();
     public Optional<Student> getStudent(int id) {
+        if(id<=0){
+            throw new IllegalArgumentException("Illegal arguments");
+        }
         return demoDao.get(id);
     }
 
@@ -23,17 +24,26 @@ public class DemoService {
     }
 
     public int postStudent(Student student) {
+        if(student==null || student.getId()<=0){
+            throw new IllegalArgumentException("Illegal arguments");
+        }
         return demoDao.post(student);
     }
 
     public int putStudent(Student student)
     {
+        if(student==null || student.getId()<=0){
+            throw new IllegalArgumentException("Illegal arguments");
+        }
         return demoDao.put(student);
     }
 
 
     public int deleteStudent(int id)
     {
+        if(id<=0){
+            throw new IllegalArgumentException("Illegal arguments");
+        }
         return demoDao.delete(id);
     }
 }
